PO-77 - made pagination workpiece.

+++ webapp/src/main/java/priceobserver/controller/ProductManageController.java
@@ -65,7 +65,7 @@ public class ProductManageController {
     }
 
     private void prepareModel(ProductTypeEnum type, Model model) {
-        model.addAttribute("pageCount", Math.ceil(productService.getProductCountByType(type) / 9.0));
+        model.addAttribute("pageCount", (int) Math.ceil(productService.getProductCountByType(type) / 9.0));
         model.addAttribute("products", productService.getProductsPageableByType(type, 0, 9));
     }
